Player + Jotai no longer singletons
