+++ src/app/providers.tsx
@@ -2,28 +2,19 @@
 import { PropsWithChildren } from 'react'
 import { TooltipProvider } from '@radix-ui/react-tooltip'
 import { Provider as JotaiProvider } from 'jotai'
-import { isBrowser } from '@/utils'
-import { SWRConfig } from 'swr'
+import { PlayerProvider } from '@/features/player'
 
 // By default, Jotai uses a single global store. This creates problems during
 // Next.js SSR where state could accidentally leak between requests for
-// different clients. Therefore, on server-only we render an extra Provider
-// which creates a store for just that request. See docs for details:
-// https://jotai.org/docs/guides/nextjs
-function MaybeJotaiProvider({ children }: PropsWithChildren<{}>) {
-  // if (isBrowser()) {
-  // TODO: maybe get rid of singletons so this actually makes sense?
-  if (true) {
-    return children
-  }
-  return <JotaiProvider>{children}</JotaiProvider>
-}
+// different clients. <JotaiProvider> below creates a tree-scoped jotai store.
+// See https://jotai.org/docs/guides/nextjs
 
 export function Providers({ children }: PropsWithChildren<{}>) {
-  // TooltipProvider should accept ReactNode...strange that it doesn't.
   return (
-    <MaybeJotaiProvider>
-      <TooltipProvider>{children as any}</TooltipProvider>
-    </MaybeJotaiProvider>
+    <JotaiProvider>
+      <PlayerProvider>
+        <TooltipProvider>{children}</TooltipProvider>
+      </PlayerProvider>
+    </JotaiProvider>
   )
 }
