+++ src/features/SongVisualization/SongVisualizer.tsx
@@ -5,6 +5,8 @@ import { useRef, useMemo } from 'react'
 import * as touchscroll from '@/features/SongVisualization/touchscroll'
 import { PIXELS_PER_SECOND as pps } from './utils'
 import { Canvas } from '@/components'
+import { usePlayer } from '../player'
+import { useStore } from 'jotai'
 
 type HandSettings = {
   [trackId: string]: {
@@ -37,6 +39,8 @@ function CanvasRenderer({
 }: CanvasRendererProps) {
   const { width, height, measureRef } = useSize()
   const canvasRef = useRef<HTMLCanvasElement>()
+  const player = usePlayer()
+  const store = useStore()
 
   const canvasRect: DOMRect = useMemo(() => {
     return canvasRef.current?.getBoundingClientRect() ?? {}
@@ -74,9 +78,9 @@ function CanvasRenderer({
     <div
       className="absolute w-full h-full touch-none"
       ref={measureRef}
-      onPointerMove={(e) => enableTouchscroll && touchscroll.handleMove(e.nativeEvent)}
-      onPointerDown={(e) => enableTouchscroll && touchscroll.handleDown(e.nativeEvent)}
-      onPointerUp={(e) => enableTouchscroll && touchscroll.handleUp(e.nativeEvent)}
+      onPointerMove={(e) => enableTouchscroll && touchscroll.handleMove(player, e.nativeEvent)}
+      onPointerDown={(e) => enableTouchscroll && touchscroll.handleDown(player, e.nativeEvent)}
+      onPointerUp={(e) => enableTouchscroll && touchscroll.handleUp(store, player, e.nativeEvent)}
     >
       <Canvas ref={canvasRef} render={renderCanvas} />
     </div>
